refactor: move out shortcut logic to its own file
