+++ go/vt/vtgate/planbuilder/testdata/onecase.txt
@@ -1,19 +1,1 @@
 # Add your test case here for debugging and run go test -run=One.
-# SelectDBA with uncorrelated subqueries
-"select t.table_schema from information_schema.tables as t where t.table_schema in (select c.column_name from information_schema.columns as c)"
-{
-  "QueryType": "SELECT",
-  "Original": "select t.table_schema from information_schema.tables as t where t.table_schema in (select c.column_name from information_schema.columns as c)",
-  "Instructions": {
-    "OperatorType": "Route",
-    "Variant": "SelectDBA",
-    "Keyspace": {
-      "Name": "main",
-      "Sharded": false
-    },
-    "FieldQuery": "select t.table_schema from information_schema.`tables` as t where 1 != 1",
-    "Query": "select t.table_schema from information_schema.`tables` as t where t.table_schema in (select c.column_name from information_schema.`columns` as c)",
-    "Table": "information_schema.`tables`"
-  }
-}
-Gen4 plan same as above
